Migrate channels-selection
